+++ src/main/java/capstone/design/control_automation/event/entity/EmergencyStatus.java
@@ -1,4 +1,4 @@
-package capstone.design.control_automation.entity;
+package capstone.design.control_automation.event.entity;
 
 public enum EmergencyStatus {
     POLICE, FIRE, MEDICAL;
